moving additional UI update functions to DOMUpdates
